add 작성자만 글 수정 삭제 기능 추가

+++ src/main/java/techcourse/myblog/presentation/controller/ArticleController.java
@@ -9,56 +9,73 @@ import techcourse.myblog.application.dto.ArticleDto;
 import techcourse.myblog.application.dto.CommentDto;
 import techcourse.myblog.application.service.ArticleService;
 import techcourse.myblog.application.service.CommentService;
+import techcourse.myblog.application.service.UserService;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Controller
 public class ArticleController {
     private final ArticleService articleService;
     private final CommentService commentService;
+    private final UserService userService;
 
     @Autowired
-    public ArticleController(final ArticleService articleService, CommentService commentService) {
+    public ArticleController(final ArticleService articleService, CommentService commentService, UserService userService) {
         this.articleService = articleService;
         this.commentService = commentService;
+        this.userService = userService;
     }
 
     @PostMapping("/articles")
-    public RedirectView createArticles(ArticleDto article) {
-        Long id = articleService.save(article);
+    public RedirectView createArticles(HttpSession httpSession, ArticleDto article) {
+        String email = (String) httpSession.getAttribute("email");
+        Long id = articleService.save(article, email);
 
         return new RedirectView("/articles/" + id);
     }
 
+
+    //todo 로그인 <-> 세션 비교 중복 제거
     @GetMapping("/articles/{articleId}")
     public ModelAndView readArticlePageByArticleId(@PathVariable Long articleId) {
         ModelAndView modelAndView = new ModelAndView("/article");
         modelAndView.addObject("article", articleService.findById(articleId));
+        modelAndView.addObject("user", articleService.findAuthor(articleId));
 
         List<CommentDto> commentDtos = commentService.findAllByArticleId(articleId);
         modelAndView.addObject("comments", commentDtos);
         return modelAndView;
     }
 
-    @GetMapping("/articles/{articleId}/edit")
-    public ModelAndView readArticleEditPage(@PathVariable Long articleId) {
-        ModelAndView modelAndView = new ModelAndView("/article-edit");
-        modelAndView.addObject("article", articleService.findById(articleId));
-        modelAndView.addObject("method", "put");
-        return modelAndView;
-    }
-
     @PutMapping("/articles/{articleId}")
-    public RedirectView updateArticle(@PathVariable Long articleId, ArticleDto article) {
+    public RedirectView updateArticle(HttpSession httpSession, @PathVariable Long articleId, ArticleDto article) {
+        String email = (String) httpSession.getAttribute("email");
+        articleService.checkAuthor(articleId, email);
+
         RedirectView redirectView = new RedirectView("/articles/" + articleId);
-        articleService.modify(articleId, article);
+        articleService.modify(articleId, article, email);
         return redirectView;
     }
 
     @DeleteMapping("/articles/{articleId}")
-    public RedirectView deleteArticle(@PathVariable Long articleId) {
+    public RedirectView deleteArticle(HttpSession httpSession, @PathVariable Long articleId) {
+        String email = (String) httpSession.getAttribute("email");
+        articleService.checkAuthor(articleId, email);
+
         RedirectView redirectView = new RedirectView("/");
         articleService.removeById(articleId);
         return redirectView;
+    }
+
+    @GetMapping("/articles/{articleId}/edit")
+    public ModelAndView readArticleEditPage(HttpSession httpSession, @PathVariable Long articleId) {
+        String email = (String) httpSession.getAttribute("email");
+        articleService.checkAuthor(articleId, email);
+
+        ModelAndView modelAndView = new ModelAndView("/article-edit");
+        modelAndView.addObject("article", articleService.findById(articleId));
+        modelAndView.addObject("method", "put");
+        return modelAndView;
     }
 }
